core: move contract-related tests to the contract package
